requirements.txt
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
boto3==1.35.2
botocore==1.35.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.8.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
dnspython==2.6.1
duckduckgo_search==6.2.10
email_validator==2.2.0
executing==2.0.1
fastapi==0.112.1
fastapi-cli==0.0.5
fastavro==1.9.5
filelock==3.15.4
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.5.2
nest-asyncio==1.6.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
primp==0.6.1
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.2
streamlit==1.37.1
tenacity==8.5.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.4
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.23.0
wcwidth==0.2.13
websockets==13.0
win32-setctime==1.1.0
